pyproject.toml
[project]
name = "zedny-agent-demo"
version = "0.1.0"
description = "Production-ready multimodal chat for Gemma-4-31B (Modal + FastAPI + Streamlit)"
requires-python = ">=3.11"
dependencies = [
    "modal>=0.64.0",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.2.0",
    "python-multipart>=0.0.9",
    "httpx>=0.27.0",
    "pyyaml>=6.0",
    "slowapi>=0.1.9",
    "structlog>=24.1.0",
    "streamlit>=1.33.0",
    "pillow>=10.2.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.4.0",
]

[tool.ruff]
line-length = 110
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = ["E501"]
